+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ## Master
 
 - Update comment instead deleting, if it has replies (BitBucket Server) [@langovoi][]
+- Add support of paged APIs of BitBucket Server [@langovoi][]
 
 # 6.1.6